Update WseEnablingStatus.ps1
What: If there is only one object returned from the Select-String results based on Dxdiag info, we can access the element directly.

Why: If only one object is returned from the Select-String results, it will automatically convert into a string object instead of a MatchInfo object array. Accessing an element on a string object will result in a character object, leading to a failed access attempt.