[MEP Automation] to leverage 'FriendlyName' from DxDiag for e2e testing (#22)
What: In the WseEnablingStatus function, we have retrieved the 'FriendlyName' of the validated camera. We can use this information to directly access the camera in the Settings page.

Why: We are not able to enumerate all the friendly names of OEM cameras, and DxDiag to provide a way to overcome this.