Initial version of CameraReliabilityTest
What: Perform a check on continuously switching between photo mode and video mode in the Camera app to ensure that the MEP function remains active and that camera streaming is still available.

Why: To assist silicon and OEM partners in ensuring the reliability of their camera drivers and to maintain the quality of MEP.

How: To run the script, use the command '.\CameraReliabilityTest.ps1 -iteration <integer>'. For example, '.\CameraReliabilityTest.ps1 -iteration 10' will switch between photo and video modes 10 times. If the '-iteration' parameter is not specified by the user, its default value is 10.

Tested how: Executed CameraReliabilityTest.ps1 on Qualcomm/Intel/AMD platforms.